BallController

Description

Documentation

ballController

Arguments

:: Ball
Initial ball

-> GameInfo
Initial game state

-> SF (GameInput, ObjsMap) GameOutput

ballController is responsible for controlling ball state. It's one big SF which uses many other SFs defined in this module.

clickSpeedModifier :: SF (Event (Point V2 Double, Point V2 Double, Double)) (V2 Double)

Our clicks should only modify velocity, based on ball position, player position and power given

hitSAModifier :: SF (Event ((GameObjType, HitDir), (V2 Double, V2 Double, Bool))) (V2 Double, V2 Double)

hitSAModifier is a Hit Speed Acceleration Modifier. On collision we will change speed and acceleration. For example, when we're rolling and friction works and we hit a wall, friction direction should change.

powerCalc

Arguments

:: Double
Speed of power change

-> Double
Initial power value (used for switching)

-> SF Bool Double

powerCalc is supposed to calculate power of ball shot. it just applies acc value to integral and switches, so we go back and forward between 0 and 100 power value

computePrevVals :: SF (Double, Double, [V2 Double]) [V2 Double]

computePrevVals updates our list of 3 recent positions

getVA

Arguments

:: Ball
Ball data to use

-> SF (V2 Double, V2 Double, Bool) (V2 Double, V2 Double)
Resulting SF counting velocity and passing acceleration

getVA this is the most important SF: it is concerned with velocity. Our acceleration values are modified discretely, and our position value is based only on velocity. Velocity values are changed in continuous and discrete manner (it changes all the time based on acceleration, and it may change according to events). Discrete changes are implemented as rSwitch in ballController, and continuous changes are implemented here.

verifyLoseCondition :: SF (Event (GameObjType, HitDir)) (Event ())

Check whether we collided with spikes

verifyWinCondition :: SF (Event (GameObjType, HitDir)) (Event ())

Check whether we collided with goal

filterFarObjectsSF :: GameInfo -> SF (ObjsMap, Point V2 Int) ObjsMap

This SF filters far objects from ObjsMap given in input

collisionCalculator

Arguments

:: (Int, Int)
Tile width and height

-> SF (ObjsMap, Point V2 Double, Double) CollisionEffect
SF to count CollisionEffect

collisionCalculator processes all objects from given ObjsMap and checks every edge for a collision with current player position

collisionChecker :: GameInfo -> SF (Point V2 Double, V2 Double, Double, ObjsMap) (Event (GameObjType, HitDir), Bool)

collisionChecker is responsible for generating informations on collisions. It uses collisionCalculator to calculate collisions. Beside generating event on collision, it gives us continuous info whether we're rolling