Game

Description

Documentation

main :: IO ()

Main is supposed to initialize all IO (create SDL Window, create Renderer, load maps and scores from files) and then start reactimate loop

appLoop

Arguments

:: Renderer
Renderer used

-> Texture
Sprite sheet

-> [[Int]]
Highscores for each level

-> [FilePath]
Filepaths for scores to write to

-> [ObjsMap]
Map of game objects with positions as keys

-> Font
Font used to write game info on screen

-> GameOutput
Yampa ouput which gives us gamestate info

-> IO Bool
Should we finish our loop?

appLoop is responsible for displaying game state with SDL and managing IO when needed (for example: saving score)