GameController

Description

Documentation

dynamicObjectController

Arguments

:: DynamicGameObj
Our object data

-> Double
Position value to add to initial position (used in switching)

-> SF () (V2 Double, GameObj)
SF which controls our object position. It returns data necessary for Ball to consider (position and type)

This simple controller is used to control the state of dynamic game objects beside the player. These objects have very basic functionanlity: initial position, positon limit, directory and speed. They just slide back and forward.

gameController

Arguments

:: Ball
Initial ball state

-> GameInfo
Initial game info

-> GameSF
Signal function from GameInput to GameOutput we switch on level change

This controller is controlling state of the whole game. For now it just gets position of all dynamic objects and uses ballController with given objects as an input to produce output