ObjectsLoader

Description

Documentation

prepareGidsToObjsMap :: GidsToObjsMap

Make a GidsToObjsMap for all objects which should be considered when loading map

prepareStaticObjsMapSingle

Arguments

:: (CInt, CInt)
tileWidth and tileHeight

-> FilePath
Map file path

-> IO ObjsMap
Loaded ObjsMap

Load a map from file and initialize ObjsMap

prepareStaticObjsMap

Arguments

:: (CInt, CInt)
tileWidth and tileHeight

-> FilePath
Directory with maps (only)

-> IO [ObjsMap]
Collection of loaded ObjsMaps

Load all maps from directory and initialize collection of ObjsMap

prepareDynamicObjs

Arguments

:: Int

-> IO [[DynamicGameObj]]
Collection of collections of DynamicGameObjs (one list for each level)

TODO load dynamic game objects. This is NOT finished: we don't load anything from files. Some objects are hard-coded just to show it works.

prepareScoresFiles

Arguments

:: FilePath
Directory where scores should be stored

-> IO [FilePath]
File paths of all scores files

Given directory where maps are, this function creates missing scores files in current directory and returns list of paths to files with scores