Types

Description

Documentation

type GameSF = SF GameInput GameOutput

data Texture

Simple wrapper around SDL Texture which also contains texture size

Constructors

Texture

Fields

texture :: Texture

SDL Texture

size :: V2 CInt

Texture size

data Ball

Necessary Ball information

Constructors

Ball

Fields

position :: Point V2 Double

Current Ball position

velocity :: V2 Double

Current Ball velocity

acceleration :: V2 Double

Current Ball acceleration

radius :: Double

Ball radius

power :: Double

Power controls how fast our shot power loads

data GameObjType

Type of Game Object

Constructors

Player

Wall

Spikes

Goal

data GameObj

GameObj contains necessary data to draw object: its type and texture coordinates in sheet

Constructors

GameObj

Fields

objType :: GameObjType

textureCoords :: Rectangle CInt

data DynamicGameObj

DynamicGameObj contains necessary data ato control dynamic object changes

Constructors

DynamicGameObj

Fields

gameObj :: GameObj

Data needed to draw object

startPos :: Point V2 Double

Starting position of object

direction :: Direction

Direction of movement

speed :: Double

Object speed

limit :: Double

Maximum deviation from start position

type GidsToObjsMap = Map Word32 GameObj

Map from tilemap Gids to game objects with corresponding textures

type ObjsMap = Map (Int, Int) GameObj

Map from window positions to gameobjs which should be drawn in corresponding positions

data GameInput

GameInput is given to main game controller

Constructors

GameInput

Fields

mPressed :: Bool

Checks whether LMB is pressed

mEventPressed :: Event ()

Fires when LMB is pressed

mEventReleased :: Event ()

Fires when LMB is released

mPos :: Point V2 CInt

Mouse position

qClick :: Event ()

Fires when Q is clicked (check whether game should end)

currTimeIn :: Integer

Passes start time to output

data GameOutput

GameOutput is returned by main game controller

Constructors

GameOutput

Fields

ball :: Ball

Output ball info

shouldEnd :: Bool

Did we click Q? Then SDL should finish process

nShots :: Int

Number of player shots

didWin :: Bool

Did player finish lvl?

didFinishGame :: Bool

Did player finish the whole game?

outLvl :: Int

Simply pass current level to output

objsMap :: ObjsMap

Pass all game objects info

currTimeOut :: Integer

Pass start time to SDL

data GameInfo

GameInfo contains initial game data

Constructors

GameInfo

Fields

screenWidth :: CInt

Window width in pixels

screenHeight :: CInt

Window height in pixels

tileWidth :: CInt

Tile width in pixels

tileHeight :: CInt

Tile height in pixels

staticObjsMap :: [ObjsMap]

Loaded static objects map for each map

dynamicObjsInfo :: [[DynamicGameObj]]

Loaded dynamic objects collection for each map

currLvl :: Int

Starting level

data CollisionEffect

Used in collisions. Contains information about side we hit and type of object we collided with on given side

Constructors

CollisionEffect

Fields

leftHitEffect :: Maybe GameObjType

rightHitEffect :: Maybe GameObjType

upHitEffect :: Maybe GameObjType

downHitEffect :: Maybe GameObjType

leftUpHitEffect :: Maybe GameObjType

rightUpHitEffect :: Maybe GameObjType

leftDownHitEffect :: Maybe GameObjType

rightDownHitEffect :: Maybe GameObjType

data HitDir

Which side of a rectangle did we hit?

Constructors

LeftSide

RightSide

UpSide

DownSide

LeftUpSide

RightUpSide

LeftDownSide

RightDownSide

data Direction

Direction of dynamic object movement

Constructors

VerticalUp

VerticalDown

HorizontalLeft

HorizontalRight